+++ NPCs/Bosses/AH/Haruka/Haruka.cs
@@ -677,7 +677,7 @@ namespace AAMod.NPCs.Bosses.AH.Haruka
                     }
                     if(strikebackproj != 0)
                     {
-                        startAngle -= .1d * (int)(strikebackproj/2);
+                        startAngle -= .1d * (strikebackproj / 2);
                         deltaAngle = spread / 3f;
                         for (int i = 0; i < strikebackproj; i++)
                         {
